OBDX-13 removed method into Notification service
