+++ grails-app/services/org/pih/warehouse/product/ProductService.groovy
@@ -35,8 +35,6 @@ class ProductService {
     def identifierService
     def userService
     def dataService
-    def productAvailabilityService
-    def notificationService
 
     def getNdcResults(operation, q) {
         def hipaaspaceApiKey = grailsApplication.config.hipaaspace.api.key
@@ -1362,36 +1360,6 @@ class ProductService {
         return productInstance
     }
 
-    void sendProductAvailabilityMessages(String locationId, List productIds){
-        Location location = Location.get(locationId)
-        List<Product> products = Product.findAllByIdInList(productIds)
-        log.info "location:${location} from locationId:${locationId}, products:${products} from ids:${productIds}"
-        Map<Product, Integer> quantityMap = productAvailabilityService.getQuantityOnHandByProduct(location, products)
-        Map<Product, Integer> quantityAvailableMap = productAvailabilityService.getQuantityAvailableToPromiseByProduct(location, products)
-        List availableItems = productAvailabilityService.getAvailableItems(location, products)
-        sendProductAvailabilityMessages(location, products, quantityMap, quantityAvailableMap, availableItems)
-    }
-
-    void sendProductAvailabilityMessages(Location location, List<Product> products, Map<Product, Integer> quantityMap,
-                                         Map<Product, Integer> quantityAvailableMap, List availableItems){
-        String TOPIC_ARN = grailsApplication.config.awssdk.sns.productAvailability
-        Attribute externalIdAttribute = findExternalProductIdAttribute()
-        JSONArray jsonArray = new JSONArray()
-        products?.each {Product product ->
-            JSONObject jsonObject = new JSONObject()
-            String externalId = product.attributes.find{it.attribute == externalIdAttribute}.value
-            List productAvailableItem = availableItems.findAll {it.product == product}
-            jsonObject.put("productId", externalId)
-            jsonObject.put("productCode", product?.productCode)
-            jsonObject.put("locationNumber", location?.locationNumber)
-            jsonObject.put("quantityOnHand", quantityMap[product])
-            jsonObject.put("quantityAvailable", quantityAvailableMap[product])
-            jsonObject.put("earliestExpirationDate", productAvailableItem?.min {it?.inventoryItem?.expirationDate})
-            log.info "JSONObject:${jsonObject.toString(2)}, for product:${product}"
-            jsonArray.put(jsonObject)
-        }
-        notificationService.publish(TOPIC_ARN, jsonArray.toString(), "Product Availability")
-    }
 
 
 }
